@@ -1,7 +1,7 @@
 import { AngularMaterialModuleModule } from './utils/angular-material-module/angular-material-module.module';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { MarkdownModule, MarkedOptions } from 'ngx-markdown';
 import { NgModule } from '@angular/core';
 
@@ -22,10 +22,9 @@ import { TddServerService} from './utils/tdd-server.service';
     AppRoutingModule,
     BrowserAnimationsModule,
     BrowserModule,
-    HttpClient,
     HttpClientModule,
     MarkdownModule.forRoot({
-      loader: HttpClient, // optional, only if you use [src] attribute
+      // loader: HttpClient, // optional, only if you use [src] attribute
       markedOptions: {
       provide: MarkedOptions,
       useValue: {
